% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{downrefs}
\alias{downrefs}
\title{table of references stored in clustifyrdata}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 6 columns.
}
\source{
various packages
}
\usage{
downrefs
}
\description{
table of references stored in clustifyrdata
}
\concept{data}
\keyword{datasets}
