% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{immgen_ref}
\alias{immgen_ref}
\title{Matrix of average gene expression from immgen data}
\format{A read count average matrix

\describe{
}}
\source{
["SingleR"]
}
\usage{
immgen_ref
}
\description{
253 mouse immune cell types
}
\keyword{datasets}
