% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCA_lung_mat}
\alias{MCA_lung_mat}
\title{Matrix of single-cell RNA-seq cells of the lung.}
\format{A sparseMatrix with genes as rows and cells as columns.

\describe{
}}
\source{
[google.com]
}
\usage{
MCA_lung_mat
}
\description{
A sample of lung cells, from mouse cell atlas project
}
\keyword{datasets}
