% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCA_lung_meta}
\alias{MCA_lung_meta}
\title{Meta-data for single-cell RNA-seq of the lung.}
\format{A data frame with 5 variables:
\describe{
}}
\source{
[google.com]
}
\usage{
MCA_lung_meta
}
\description{
A sample of lung cells, from mouse cell atlas project
}
\keyword{datasets}
