% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse.rnaseq_ref}
\alias{mouse.rnaseq_ref}
\title{Matrix of average gene expression from general mouse cell type data}
\format{A read count matrix

\describe{
}}
\source{
["SingleR"]
}
\usage{
mouse.rnaseq_ref
}
\description{
28 general cell types
}
\keyword{datasets}
