% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_indrop_avg}
\alias{pan_indrop_avg}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq pancreas data using indrop method.}
\format{A sparseMatrix with genes as rows and clusters as columns.

\describe{
}}
\source{
["hemberg"]
}
\usage{
pan_indrop_avg
}
\description{
A sample of 8569 pancreas cells
}
\keyword{datasets}
