% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_indrop_meta}
\alias{pan_indrop_meta}
\title{Meta-data for single-cell RNA-seq pancreas data using indrop method.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8569 rows and 5 columns.
}
\source{
["hemberg"]
}
\usage{
pan_indrop_meta
}
\description{
A sample of 8569 pancreas cells
}
\keyword{datasets}
