% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_indrop_vargenes}
\alias{pan_indrop_vargenes}
\title{variable genes identified by Seurat from single-cell RNA-seq pancreas data using indrop method.}
\format{
An object of class \code{character} of length 2788.
}
\source{
["hemberg"]
}
\usage{
pan_indrop_vargenes
}
\description{
A sample of 8569 pancreas cells
}
\keyword{datasets}
