% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_smartseq2_avg}
\alias{pan_smartseq2_avg}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq pancreas data using smartseq2 method.}
\format{A matrix with genes as rows and clusters as columns.

\describe{
}}
\source{
["hemberg"]
}
\usage{
pan_smartseq2_avg
}
\description{
A sample of 2209 pancreas cells
}
\keyword{datasets}
