% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_smartseq2_matrix}
\alias{pan_smartseq2_matrix}
\title{Matrix of single-cell RNA-seq pancreas data using smartseq2 method.}
\format{
An object of class \code{dgCMatrix} with 25525 rows and 2209 columns.
}
\source{
["hemberg"]
}
\usage{
pan_smartseq2_matrix
}
\description{
A sample of 2209 pancreas cells
}
\keyword{datasets}
