% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_smartseq2_meta}
\alias{pan_smartseq2_meta}
\title{Meta-data for single-cell RNA-seq pancreas data using smartseq2 method.}
\format{A data frame with 4 variables:
\describe{
}}
\source{
["hemberg"]
}
\usage{
pan_smartseq2_meta
}
\description{
A sample of 2209 pancreas cells
}
\keyword{datasets}
