% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_smartseq2_vargenes}
\alias{pan_smartseq2_vargenes}
\title{variable genes identified by Seurat from single-cell RNA-seq pancreas data using smartseq2 method.}
\format{A vector:
\describe{
}}
\source{
["hemberg"]
}
\usage{
pan_smartseq2_vargenes
}
\description{
A sample of 2209 pancreas cells
}
\keyword{datasets}
