% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc5_markers}
\alias{pbmc5_markers}
\title{Marker genes identified by Seurat from single-cell RNA-seq PBMCs from 5'end kit.}
\format{
An object of class \code{data.frame} with 5926 rows and 7 columns.
}
\source{
[google.com]
}
\usage{
pbmc5_markers
}
\description{
A sample of 8000 cells from a bigger experiment
}
\keyword{datasets}
