% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc5_meta}
\alias{pbmc5_meta}
\title{Meta-data for single-cell RNA-seq PBMCs from 5'end kit.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8555 rows and 4 columns.
}
\source{
<https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k>
}
\usage{
pbmc5_meta
}
\description{
A sample of 8000 cells from a bigger experiment
}
\seealso{
Other 10x: 
\code{\link{pbmc5_markers}},
\code{\link{pbmc5_matrix}},
\code{\link{pbmc_avg}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}}
}
\concept{10x}
\keyword{datasets}
