% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_avg}
\alias{pbmc_avg}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq PBMCs.}
\format{A matrix with genes as rows and clusters as columns.

\describe{
}}
\source{
[google.com]
}
\usage{
pbmc_avg
}
\description{
A sample of 3k PBMC cells
}
\keyword{datasets}
