% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers}
\alias{pbmc_markers}
\title{Marker genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{A data frame with 7 variables:
\describe{
}}
\source{
[google.com]
}
\usage{
pbmc_markers
}
\description{
A sample of 3k cells
}
\keyword{datasets}
