% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers_M3Drop}
\alias{pbmc_markers_M3Drop}
\title{Marker genes identified by M3Drop from single-cell RNA-seq PBMCs.}
\format{
An object of class \code{data.frame} with 162 rows and 3 columns.
}
\source{
[google.com]
}
\usage{
pbmc_markers_M3Drop
}
\description{
A sample of 3k cells
}
\keyword{datasets}
