% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_matrix}
\alias{pbmc_matrix}
\title{Matrix of single-cell RNA-seq PBMCs.}
\format{
An object of class \code{dgCMatrix} with 13714 rows and 2638 columns.
}
\source{
[google.com]
}
\usage{
pbmc_matrix
}
\description{
A sample of 3k PBMC cells
}
\keyword{datasets}
