% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_vargenes}
\alias{pbmc_vargenes}
\title{variable genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{
An object of class \code{character} of length 2000.
}
\source{
[google.com]
}
\usage{
pbmc_vargenes
}
\description{
A sample of 3k cells
}
\keyword{datasets}
