% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_moca_main}
\alias{ref_moca_main}
\title{Matrix of average gene expression per cluster from single-cell sci-RNA-seq3 from Mouse Organogenesis Cell Atlas.}
\format{
An object of class \code{matrix} with 26183 rows and 37 columns.
}
\source{
<http://atlas.gs.washington.edu/mouse-rna/>
}
\usage{
ref_moca_main
}
\description{
100,000 cells
}
\keyword{datasets}
