% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_tabula_muris_facs}
\alias{ref_tabula_muris_facs}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq from Tabula Muris (SMART-Seq2).}
\format{
An object of class \code{matrix} with 23341 rows and 175 columns.
}
\source{
<https://tabula-muris.ds.czbiohub.org/>
}
\usage{
ref_tabula_muris_facs
}
\description{
53,760 cells from 20 tissues from 8 mice
}
\keyword{datasets}
