% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_small3}
\alias{s_small3}
\title{Small clustered Seurat3 object}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\source{
`pbmc_small` processed by `[Seurat]`
}
\usage{
s_small3
}
\description{
Small clustered Seurat3 object
}
\seealso{
Other obj: 
\code{\link{s_small}}
}
\concept{obj}
\keyword{datasets}
