% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yan_ref}
\alias{yan_ref}
\title{Matrix of average gene expression from `yan` human embryo data}
\format{A read count matrix

\describe{
}}
\source{
["scmap"]
}
\usage{
yan_ref
}
\description{
6 early development cell types
}
\keyword{datasets}
