% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_bulk_matrix}
\alias{gtex_bulk_matrix}
\title{Matrix of average gene expression from GTEx recount bulk RNA seq.}
\format{
An object of class \code{matrix} with 25224 rows and 54 columns.
}
\source{
["SRP012682"]
}
\usage{
gtex_bulk_matrix
}
\description{
53,760 cells from 20 tissues from 8 mice
}
\keyword{datasets}
