% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtex_bulk_matrix}
\alias{gtex_bulk_matrix}
\title{Matrix of average gene expression from GTEx recount bulk RNA seq.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 25224 rows and 54 columns.
}
\source{
<https://www.ebi.ac.uk/ena/data/view/PRJNA75899>
}
\usage{
gtex_bulk_matrix
}
\description{
53,760 cells from 20 tissues from 8 mice
}
\seealso{
Other ref: 
\code{\link{ref_MCA}},
\code{\link{ref_cortex_dev}},
\code{\link{ref_hema_microarray}},
\code{\link{ref_immgen}},
\code{\link{ref_moca_main}},
\code{\link{ref_mouse.rnaseq}},
\code{\link{ref_pan_indrop}},
\code{\link{ref_pan_smartseq2}},
\code{\link{ref_tabula_muris_drop}},
\code{\link{ref_tabula_muris_facs}}
}
\concept{ref}
\keyword{datasets}
