% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCA_lung_mat}
\alias{MCA_lung_mat}
\title{Matrix of single-cell RNA-seq cells of the lung.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 17097 rows and 6940 columns.
}
\source{
<https://figshare.com/s/865e694ad06d5857db4b>
}
\usage{
MCA_lung_mat
}
\description{
A sample of lung cells, from mouse cell atlas project
}
\seealso{
Other MCA: 
\code{\link{MCA_lung_meta}}
}
\concept{MCA}
\keyword{datasets}
