% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MCA_lung_meta}
\alias{MCA_lung_meta}
\title{Meta-data for single-cell RNA-seq of the lung.}
\format{
An object of class \code{data.frame} with 6940 rows and 6 columns.
}
\source{
<https://figshare.com/s/865e694ad06d5857db4b>
}
\usage{
MCA_lung_meta
}
\description{
A sample of lung cells, from mouse cell atlas project
}
\seealso{
Other MCA: 
\code{\link{MCA_lung_mat}}
}
\concept{MCA}
\keyword{datasets}
