% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_indrop_markers}
\alias{pan_indrop_markers}
\title{Marker genes identified by Seurat from single-cell RNA-seq pancreas data using indrop method.}
\format{
An object of class \code{data.frame} with 7837 rows and 7 columns.
}
\source{
["hemberg"]
}
\usage{
pan_indrop_markers
}
\description{
A sample of 8569 pancreas cells
}
\keyword{datasets}
