% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_indrop_matrix}
\alias{pan_indrop_matrix}
\title{Matrix of single-cell RNA-seq pancreas data using indrop method.}
\format{
An object of class \code{dgCMatrix} with 20125 rows and 8569 columns.
}
\source{
["hemberg"]
}
\usage{
pan_indrop_matrix
}
\description{
A sample of 8569 pancreas cells
}
\keyword{datasets}
