% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_indrop_meta}
\alias{pan_indrop_meta}
\title{Meta-data for single-cell RNA-seq pancreas data using indrop method.}
\format{A data frame with 4 variables:
\describe{
}}
\source{
["hemberg"]
}
\usage{
pan_indrop_meta
}
\description{
A sample of 8569 pancreas cells
}
\keyword{datasets}
