% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_indrop_meta}
\alias{pan_indrop_meta}
\title{Meta-data for single-cell RNA-seq pancreas data using indrop method.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8569 rows and 5 columns.
}
\source{
<https://hemberg-lab.github.io/scRNA.seq.datasets/human/pancreas/>
}
\usage{
pan_indrop_meta
}
\description{
A sample of 8569 pancreas cells
}
\seealso{
Other hemberg: 
\code{\link{pan_indrop_markers}},
\code{\link{pan_indrop_matrix}},
\code{\link{pan_indrop_vargenes}},
\code{\link{pan_smartseq2_matrix}},
\code{\link{pan_smartseq2_meta}},
\code{\link{pan_smartseq2_vargenes}}
}
\concept{hemberg}
\keyword{datasets}
