% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pan_smartseq2_matrix}
\alias{pan_smartseq2_matrix}
\title{Matrix of single-cell RNA-seq pancreas data using smartseq2 method.}
\format{
An object of class \code{dgCMatrix} with 25525 rows and 2209 columns.
}
\source{
<https://hemberg-lab.github.io/scRNA.seq.datasets/human/pancreas/>
}
\usage{
pan_smartseq2_matrix
}
\description{
A sample of 2209 pancreas cells
}
\seealso{
Other hemberg: 
\code{\link{pan_indrop_markers}},
\code{\link{pan_indrop_matrix}},
\code{\link{pan_indrop_meta}},
\code{\link{pan_indrop_vargenes}},
\code{\link{pan_smartseq2_meta}},
\code{\link{pan_smartseq2_vargenes}}
}
\concept{hemberg}
\keyword{datasets}
