% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc5_markers}
\alias{pbmc5_markers}
\title{Marker genes identified by Seurat from single-cell RNA-seq PBMCs from 5'end kit.}
\format{
An object of class \code{data.frame} with 5926 rows and 7 columns.
}
\source{
<https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k>
}
\usage{
pbmc5_markers
}
\description{
A sample of 8000 cells from a bigger experiment
}
\seealso{
Other 10x: 
\code{\link{pbmc5_matrix}},
\code{\link{pbmc5_meta}},
\code{\link{pbmc_avg}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}}
}
\concept{10x}
\keyword{datasets}
