% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc5_matrix}
\alias{pbmc5_matrix}
\title{Matrix of single-cell RNA-seq PBMCs from 5'end kit.}
\format{
An object of class \code{dgCMatrix} with 18587 rows and 8555 columns.
}
\source{
[google.com]
}
\usage{
pbmc5_matrix
}
\description{
A sample of 8000 cells from a bigger experiment
}
\keyword{datasets}
