% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_avg}
\alias{pbmc_avg}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq PBMCs.}
\format{
An object of class \code{matrix} with 13714 rows and 9 columns.
}
\source{
[google.com]
}
\usage{
pbmc_avg
}
\description{
A sample of 3k PBMC cells
}
\keyword{datasets}
