% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_avg}
\alias{pbmc_avg}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq PBMCs.}
\format{
An object of class \code{matrix} with 13714 rows and 9 columns.
}
\source{
<https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k>
}
\usage{
pbmc_avg
}
\description{
A sample of 3k PBMC cells
}
\seealso{
Other 10x: 
\code{\link{pbmc5_markers}},
\code{\link{pbmc5_matrix}},
\code{\link{pbmc5_meta}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix}},
\code{\link{pbmc_meta}},
\code{\link{pbmc_vargenes}}
}
\concept{10x}
\keyword{datasets}
