% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers}
\alias{pbmc_markers}
\title{Marker genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{
An object of class \code{data.frame} with 2304 rows and 7 columns.
}
\source{
[google.com]
}
\usage{
pbmc_markers
}
\description{
A sample of 3k cells
}
\keyword{datasets}
