% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_markers_M3Drop}
\alias{pbmc_markers_M3Drop}
\title{Marker genes identified by M3Drop from single-cell RNA-seq PBMCs.}
\format{A data frame with 3 variables:
\describe{
}}
\source{
[google.com]
}
\usage{
pbmc_markers_M3Drop
}
\description{
A sample of 3k cells
}
\keyword{datasets}
