% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_matrix}
\alias{pbmc_matrix}
\title{Matrix of single-cell RNA-seq PBMCs.}
\format{A sparseMatrix with genes as rows and cells as columns.

\describe{
}}
\source{
[google.com]
}
\usage{
pbmc_matrix
}
\description{
A sample of 3k PBMC cells
}
\keyword{datasets}
