% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_meta}
\alias{pbmc_meta}
\title{Meta-data for single-cell RNA-seq PBMCs.}
\format{A data frame with 5 variables:
\describe{
}}
\source{
[google.com]
}
\usage{
pbmc_meta
}
\description{
A sample of 3k cells
}
\keyword{datasets}
