% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_vargenes}
\alias{pbmc_vargenes}
\title{variable genes identified by Seurat from single-cell RNA-seq PBMCs.}
\format{
An object of class \code{character} of length 2000.
}
\source{
<https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k>
}
\usage{
pbmc_vargenes
}
\description{
A sample of 3k cells
}
\seealso{
Other 10x: 
\code{\link{pbmc5_markers}},
\code{\link{pbmc5_matrix}},
\code{\link{pbmc5_meta}},
\code{\link{pbmc_avg}},
\code{\link{pbmc_markers_M3Drop}},
\code{\link{pbmc_markers}},
\code{\link{pbmc_matrix}},
\code{\link{pbmc_meta}}
}
\concept{10x}
\keyword{datasets}
