% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_cortex_dev}
\alias{ref_cortex_dev}
\title{Matrix of average gene detection on cortex development cell types}
\format{A matrix with genes as rows and clusters as columns.

\describe{
}}
\source{
[https://cells.ucsc.edu/?ds=cortex-dev#]
}
\usage{
ref_cortex_dev
}
\description{
47 distinct annotated clusters of human neuronal subtypes
}
\keyword{datasets}
