% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_hema_microarray}
\alias{ref_hema_microarray}
\title{Matrix of average gene detection by microarray on sorted immune cell populations}
\format{A sparseMatrix with genes as rows and cell types as columns.

\describe{
}}
\source{
[https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE24759]
}
\usage{
ref_hema_microarray
}
\description{
38 distinct purified populations of human hematopoietic cells
}
\keyword{datasets}
