% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_hema_microarray}
\alias{ref_hema_microarray}
\title{Matrix of average gene detection by microarray on sorted immune cell populations}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 21246 rows and 38 columns.
}
\source{
[https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE24759]
}
\usage{
ref_hema_microarray
}
\description{
38 distinct purified populations of human hematopoietic cells
}
\seealso{
Other ref: 
\code{\link{gtex_bulk_matrix}},
\code{\link{ref_MCA}},
\code{\link{ref_cortex_dev}},
\code{\link{ref_immgen}},
\code{\link{ref_moca_main}},
\code{\link{ref_mouse.rnaseq}},
\code{\link{ref_pan_indrop}},
\code{\link{ref_pan_smartseq2}},
\code{\link{ref_tabula_muris_drop}},
\code{\link{ref_tabula_muris_facs}}
}
\concept{ref}
\keyword{datasets}
