% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_immgen}
\alias{ref_immgen}
\title{Matrix of average gene expression from immgen data}
\format{
An object of class \code{matrix} with 22134 rows and 253 columns.
}
\source{
["SingleR"]
}
\usage{
ref_immgen
}
\description{
253 mouse immune cell types
}
\keyword{datasets}
