% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_MCA}
\alias{ref_MCA}
\title{Matrix of average gene expression for single-cell RNA-seq.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8601 rows and 713 columns.
}
\source{
<https://figshare.com/s/865e694ad06d5857db4b>
}
\usage{
ref_MCA
}
\description{
from mouse cell atlas project
}
\seealso{
Other ref: 
\code{\link{gtex_bulk_matrix}},
\code{\link{ref_cortex_dev}},
\code{\link{ref_hema_microarray}},
\code{\link{ref_immgen}},
\code{\link{ref_moca_main}},
\code{\link{ref_mouse.rnaseq}},
\code{\link{ref_pan_indrop}},
\code{\link{ref_pan_smartseq2}},
\code{\link{ref_tabula_muris_drop}},
\code{\link{ref_tabula_muris_facs}}
}
\concept{ref}
\keyword{datasets}
