% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_MCA}
\alias{ref_MCA}
\title{Matrix of average gene expression for single-cell RNA-seq.}
\format{
An object of class \code{matrix} with 8601 rows and 713 columns.
}
\source{
[google.com]
}
\usage{
ref_MCA
}
\description{
from mouse cell atlas project
}
\keyword{datasets}
