% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_moca_main}
\alias{ref_moca_main}
\title{Matrix of average gene expression per cluster from single-cell sci-RNA-seq3 from Mouse Organogenesis Cell Atlas.}
\format{A sparseMatrix with genes as rows and clusters as columns.

\describe{
}}
\source{
[http://atlas.gs.washington.edu/mouse-rna/]
}
\usage{
ref_moca_main
}
\description{
100,000 cells
}
\keyword{datasets}
