% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_mouse.rnaseq}
\alias{ref_mouse.rnaseq}
\title{Matrix of average gene expression from general mouse cell type data}
\format{
An object of class \code{matrix} with 21214 rows and 28 columns.
}
\source{
["SingleR"]
}
\usage{
ref_mouse.rnaseq
}
\description{
28 general cell types
}
\keyword{datasets}
