% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_mouse.rnaseq}
\alias{ref_mouse.rnaseq}
\title{Matrix of average gene expression from general mouse cell type data}
\format{
An object of class \code{matrix} with 21214 rows and 28 columns.
}
\source{
["SingleR"]
}
\usage{
ref_mouse.rnaseq
}
\description{
28 general cell types
}
\seealso{
Other ref: 
\code{\link{gtex_bulk_matrix}},
\code{\link{ref_MCA}},
\code{\link{ref_cortex_dev}},
\code{\link{ref_hema_microarray}},
\code{\link{ref_immgen}},
\code{\link{ref_moca_main}},
\code{\link{ref_pan_indrop}},
\code{\link{ref_pan_smartseq2}},
\code{\link{ref_tabula_muris_drop}},
\code{\link{ref_tabula_muris_facs}}
}
\concept{ref}
\keyword{datasets}
