% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_pan_indrop}
\alias{ref_pan_indrop}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq pancreas data using indrop method.}
\format{A matrix with genes as rows and clusters as columns.

\describe{
}}
\source{
["hemberg"]
}
\usage{
ref_pan_indrop
}
\description{
A sample of 8569 pancreas cells
}
\keyword{datasets}
