% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_pan_smartseq2}
\alias{ref_pan_smartseq2}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq pancreas data using smartseq2 method.}
\format{
An object of class \code{data.frame} with 25525 rows and 12 columns.
}
\source{
["hemberg"]
}
\usage{
ref_pan_smartseq2
}
\description{
A sample of 2209 pancreas cells
}
\keyword{datasets}
