% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_pan_smartseq2}
\alias{ref_pan_smartseq2}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq pancreas data using smartseq2 method.}
\format{
An object of class \code{data.frame} with 25525 rows and 12 columns.
}
\source{
<https://hemberg-lab.github.io/scRNA.seq.datasets/human/pancreas/>
}
\usage{
ref_pan_smartseq2
}
\description{
A sample of 2209 pancreas cells
}
\seealso{
Other ref: 
\code{\link{gtex_bulk_matrix}},
\code{\link{ref_MCA}},
\code{\link{ref_cortex_dev}},
\code{\link{ref_hema_microarray}},
\code{\link{ref_immgen}},
\code{\link{ref_moca_main}},
\code{\link{ref_mouse.rnaseq}},
\code{\link{ref_pan_indrop}},
\code{\link{ref_tabula_muris_drop}},
\code{\link{ref_tabula_muris_facs}}
}
\concept{ref}
\keyword{datasets}
