% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_pan_smartseq2}
\alias{ref_pan_smartseq2}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq pancreas data using smartseq2 method.}
\format{
An object of class \code{data.frame} with 25525 rows and 12 columns.
}
\source{
<https://hemberg-lab.github.io/scRNA.seq.datasets/human/pancreas/>
}
\usage{
ref_pan_smartseq2
}
\description{
A sample of 2209 pancreas cells
}
\seealso{
Other hemberg: 
\code{\link{pan_indrop_markers}},
\code{\link{pan_indrop_matrix}},
\code{\link{pan_indrop_meta}},
\code{\link{pan_indrop_vargenes}},
\code{\link{pan_smartseq2_matrix}},
\code{\link{pan_smartseq2_meta}},
\code{\link{pan_smartseq2_vargenes}},
\code{\link{ref_pan_indrop}}
}
\concept{hemberg}
\keyword{datasets}
