% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_tabula_muris_drop}
\alias{ref_tabula_muris_drop}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq from Tabula Muris (10x genomics).}
\format{A sparseMatrix with genes as rows and clusters as columns.

\describe{
}}
\source{
[https://tabula-muris.ds.czbiohub.org/]
}
\usage{
ref_tabula_muris_drop
}
\description{
422,803 droplet libraries, 55,656 of which passed a QC cutoff of 500 genes and 1000 UMI
}
\keyword{datasets}
