% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_tabula_muris_facs}
\alias{ref_tabula_muris_facs}
\title{Matrix of average gene expression per cluster from single-cell RNA-seq from Tabula Muris (SMART-Seq2).}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 23341 rows and 175 columns.
}
\source{
<https://tabula-muris.ds.czbiohub.org/>
}
\usage{
ref_tabula_muris_facs
}
\description{
53,760 cells from 20 tissues from 8 mice
}
\seealso{
Other ref: 
\code{\link{gtex_bulk_matrix}},
\code{\link{ref_MCA}},
\code{\link{ref_cortex_dev}},
\code{\link{ref_hema_microarray}},
\code{\link{ref_immgen}},
\code{\link{ref_moca_main}},
\code{\link{ref_mouse.rnaseq}},
\code{\link{ref_pan_indrop}},
\code{\link{ref_pan_smartseq2}},
\code{\link{ref_tabula_muris_drop}}
}
\concept{ref}
\keyword{datasets}
